WIP trying to refactor and encapsulate []model.ScoreUpdate modification consider nested elements like Repeat, OnRepetitions, Chord, etc.
