+++ client/interop/musicxml/importer/importer.go
@@ -7,8 +7,6 @@ import (
 	"github.com/beevik/etree"
 	"github.com/logrusorgru/aurora"
 	"io"
-	"reflect"
-	"sort"
 )
 
 // musicXMLImporter contains global state for importing a MusicXML file
@@ -67,146 +65,6 @@ func newMusicXMLVoice() *musicXMLVoice {
 		slurs:       0, // By default a note is not slurred
 	}
 }
-
-// generateScoreUpdates flattens score updates in an importer to a single slice
-// This represents the overall output of the importer
-func (importer *musicXMLImporter) generateScoreUpdates() []model.ScoreUpdate {
-	var updates []model.ScoreUpdate
-
-	// We process parts in order of ID
-	var partIDs []string
-	for id := range importer.parts {
-		partIDs = append(partIDs, id)
-	}
-
-	sort.Sort(sort.StringSlice(partIDs))
-
-	for _, id := range partIDs {
-		part := importer.parts[id]
-		partDeclaration := model.PartDeclaration{
-			Names: part.instruments,
-		}
-		updates = append(updates, partDeclaration)
-
-		if len(part.voices) == 1 {
-			// For a single voice, we don't include a voice marker
-			updates = append(updates, part.currentVoice.updates...)
-		} else {
-			// Process voices in order of voice number
-			var voiceNumber int32 = 1
-			for voicesLeft := len(part.voices); voicesLeft > 0; voiceNumber++ {
-				if voice, ok := part.voices[voiceNumber]; ok {
-					voiceMarker := model.VoiceMarker{
-						VoiceNumber: voiceNumber,
-					}
-					updates = append(updates, voiceMarker)
-					updates = append(updates, voice.updates...)
-					voicesLeft--
-				}
-			}
-		}
-	}
-
-	return updates
-}
-
-// updates returns the current list of model.ScoreUpdate to import into
-func (importer *musicXMLImporter) updates() []model.ScoreUpdate {
-	voiceUpdates := importer.currentPart.currentVoice.updates
-
-	if eventSequence, ok := getLastRepeatEventSequence(voiceUpdates); ok {
-		return eventSequence.Events
-	} else {
-		return voiceUpdates
-	}
-}
-
-// append appends an update to the current voice
-// append increments both part-level and voice-level beats
-func (importer *musicXMLImporter) append(newUpdates ...model.ScoreUpdate) {
-	voiceUpdates := importer.currentPart.currentVoice.updates
-
-	beats := getBeats(newUpdates...)
-	importer.currentPart.beats += beats
-	importer.currentPart.currentVoice.beats += beats
-
-	if eventSequence, ok := getLastRepeatEventSequence(voiceUpdates); ok {
-		eventSequence.Events = append(
-			eventSequence.Events,
-			newUpdates...
-		)
-	} else {
-		importer.currentPart.currentVoice.updates = append(
-			importer.currentPart.currentVoice.updates,
-			newUpdates...,
-		)
-	}
-}
-
-// set sets the updates for the current voice
-// set recomputes both part-level and voice-level beats
-func (importer *musicXMLImporter) set(newUpdates []model.ScoreUpdate) {
-	voiceUpdates := importer.currentPart.currentVoice.updates
-
-	beats := getBeats(newUpdates...)
-	importer.currentPart.beats = beats
-	importer.currentPart.currentVoice.beats = beats
-
-	if eventSequence, ok := getLastRepeatEventSequence(voiceUpdates); ok {
-		eventSequence.Events = newUpdates
-	} else {
-		importer.currentPart.currentVoice.updates = newUpdates
-	}
-}
-
-// Identifies if the last update is a repeat, and returns its event sequence
-func getLastRepeatEventSequence(
-	updates []model.ScoreUpdate,
-) (*model.EventSequence, bool) {
-	if len(updates) > 0 {
-		switch value := updates[len(updates) - 1].(type) {
-		case model.Repeat:
-			// We can cast to an event sequence because in our importer we only
-			// ever repeat sequences, never individual notes
-			eventSequence := value.Event.(model.EventSequence)
-			return &eventSequence, true
-		}
-	}
-	return nil, false
-}
-
-// getBeats counts beats for a slice of model.ScoreUpdate
-func getBeats(updates ...model.ScoreUpdate) float64 {
-	beats := 0.0
-	for _, update := range updates {
-		switch value := update.(type) {
-		case model.Note:
-			beats += value.Duration.Beats()
-		case model.Rest:
-			beats += value.Duration.Beats()
-		case model.Chord:
-			min := 0.0
-			for _, event := range value.Events {
-				eventBeats := getBeats(event)
-				if eventBeats < min {
-					min = eventBeats
-				}
-			}
-			beats += min
-		case model.Repeat:
-			eventSequence := value.Event.(model.EventSequence)
-			beats += getBeats(eventSequence.Events...)
-		}
-	}
-	return beats
-}
-
-var noteType = reflect.TypeOf(model.Note{})
-var chordType = reflect.TypeOf(model.Chord{})
-var restType = reflect.TypeOf(model.Rest{})
-var attributeUpdateType = reflect.TypeOf(model.AttributeUpdate{})
-var octaveUpType = reflect.TypeOf(model.OctaveUp{})
-var octaveDownType = reflect.TypeOf(model.OctaveDown{})
 
 func warnWhileParsing(element *etree.Element, message string) {
 	log.Warn().
